Rename `themed_bottom_container` to `themed_container`
