+++ lib/application/pages/details/collection_details_page.dart
@@ -9,7 +9,7 @@ import 'package:memo/application/theme/theme_controller.dart';
 import 'package:memo/application/view-models/details/collection_details_vm.dart';
 import 'package:memo/application/widgets/theme/hero_collection_card.dart';
 import 'package:memo/application/widgets/theme/resources_list.dart';
-import 'package:memo/application/widgets/theme/themed_bottom_container.dart';
+import 'package:memo/application/widgets/theme/themed_container.dart';
 import 'package:memo/application/widgets/theme/themed_text_tag.dart';
 
 class CollectionDetailsPage extends HookWidget {
